phase4_backend/requirements.txt
# Phase 4: Backend RAG API
fastapi>=0.104.0
uvicorn>=0.24.0
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
